requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-Mail
python-dotenv
stripe
PyPDF2
pdf2docx
docx2pdf
img2pdf
PyMuPDF
python-docx
Pillow
gunicorn
psycopg2-binary
Werkzeug
Jinja2
python-dateutil
requests
